requirements.txt
apache-airflow==2.9.0
apache-airflow-providers-fab==1.2.0
sqlalchemy>=1.4.36,<2.0
flask>=2.2,<2.3
